Changed default id to -1 as 0 is used by HEAD joint in loops.
Rewritted 3D renderer to better apply descriptive data to rendering. However, doesn't look like it's working quite yet.

Integrated NodeFilter back into the project

+++ include/nodes/Predictor.h
@@ -20,7 +20,7 @@
 /// Based on equations in paper: P. Kalekar, “Time series forecasting using
 /// holt-winters exponential smoothing,” tech. rep., Kanwal Rekhi School of
 /// Information Technology, Mumbai, India, 2004.
-template<class T, class Scalar=float>
+template<class T, class Scalar=double>
 class Predictor
 {
 public:
